+++ core/skill.go
@@ -6,14 +6,15 @@ import (
 )
 
 type Skill struct {
-	Id           uuid.UUID `json:"id"`
-	SkillId      string    `json:"skillId" binding:"required"`
-	UserId       uuid.UUID `json:"userId"`
-	Exp          int       `json:"exp"`
-	Txp          int       `json:"txp"`
-	Level        int       `json:"level"`
-	DateCreated  time.Time `json:"dateCreated"`
-	DateModified time.Time `json:"dateModified"`
+	Id             uuid.UUID `json:"id"`
+	SkillId        string    `json:"skillId" binding:"required"`
+	UserId         uuid.UUID `json:"userId"`
+	Exp            int       `json:"exp"`
+	Txp            int       `json:"txp"`
+	Level          int       `json:"level"`
+	DateCreated    time.Time `json:"dateCreated"`
+	DateModified   time.Time `json:"dateModified"`
+	DateLastTxpAdd time.Time `json:"dateLastTxpAdd"`
 }
 
 var skillTable = createSkillTable()